requirements.txt
Django>=5.0,<5.1
Pillow>=10.0.0
django-admin-sortable2>=2.1.0
